@@ -31,15 +31,18 @@ class Image < ActiveRecord::Base
   validates_numericality_of :flickr_photo_id, if: Proc.new { |i| i.flickr_photo_id.present? }
 
   # Other properties (e.g. accepts_nested_attributes_for)
-  attr_accessible :asset, :asset_cache, :album_id, :title, :title_ua, :desc, :desc_ua, :place, :place_ua, :date,
-                  :updated_at, :published_at_checkbox, :tags, :tags_resolved, :flickr_photo_id, :flickr_comment_time,
-                  :deviantart_link, :istockphoto_link, :shutterstock_link, :is_for_sale, :image_width, :image_height
-  attr_taggable :tags
+  # TODO: move to strong_parameters
+  # attr_accessible :asset, :asset_cache, :album_id, :title, :title_ua, :desc, :desc_ua, :place, :place_ua, :date,
+  #                 :updated_at, :published_at_checkbox, :tags, :tags_resolved, :flickr_photo_id, :flickr_comment_time,
+  #                 :deviantart_link, :istockphoto_link, :shutterstock_link, :is_for_sale, :image_width, :image_height
+  # TODO: implement
+  #attr_taggable :tags
 
   # Model dictionaries, state machine
 
   # Scopes
-  default_scope order: 'images.published_at DESC, images.created_at DESC'
+  # TODO: move to controller
+  #default_scope order: 'images.published_at DESC, images.created_at DESC'
 
   scope :published, -> { where('images.published_at IS NOT NULL') }
 
