@@ -23,8 +23,9 @@ class Photo < ActiveRecord::Base
   validates_presence_of :asset, :owner_id, :owner_type
 
   # Other properties (e.g. accepts_nested_attributes_for)
-  attr_accessible :asset, :asset_cache, :owner_id, :owner_type, :desc, :desc_ua, :image_width, :image_height,
-                  :is_cover, :weight
+  # TODO: move to strong_parameters
+  # attr_accessible :asset, :asset_cache, :owner_id, :owner_type, :desc, :desc_ua, :image_width, :image_height,
+  #                 :is_cover, :weight
 
   # Model dictionaries, state machine
 
