@@ -14,8 +14,9 @@ class Project < ActiveRecord::Base
   validates_uniqueness_of :title, :title_ua
 
   # Other properties (e.g. accepts_nested_attributes_for)
-  attr_accessible :title, :title_ua, :is_published, :description, :description_ua, :weight, :info, :info_ua,
-                  :photos_attributes
+  # TODO: move to strong_parameters
+  # attr_accessible :title, :title_ua, :is_published, :description, :description_ua, :weight, :info, :info_ua,
+  #                 :photos_attributes
 
   accepts_nested_attributes_for :photos, allow_destroy: true, reject_if: proc { |a| !a.has_key?('id') && !a.has_key?('asset') }
 
