@@ -2,8 +2,9 @@ class User < ActiveRecord::Base
   # disable registrations on PROD
   devise :database_authenticatable, :recoverable, :rememberable, :registerable
 
-  # Setup accessible (or protected) attributes for your model
-  attr_accessible :email, :password, :password_confirmation, :remember_me
+  # # Setup accessible (or protected) attributes for your model
+  # TODO: move to strong_parameters
+  # attr_accessible :email, :password, :password_confirmation, :remember_me
 
   validates_presence_of :password, :password_confirmation
   validates_length_of :password, minimum: 6
