@@ -1,22 +1,18 @@
-Portfolio::Application.routes.draw do
+Rails.application.routes.draw do
   if Rails.env.production?
     devise_for :users, controllers: { registrations: 'registrations' }
   else
     devise_for :users
   end
 
-  authenticated :user do
-    root to: 'home#index'
-  end
-
-  match '/rss' => 'home#rss', as: :rss_feed, defaults: { format: 'rss' }
-  match '/contacts' => 'home#contacts', as: :contacts
-
-  root to: 'home#index'
+  get '/rss' => 'home#rss', as: :rss_feed, defaults: { format: 'rss' }
+  get '/contacts' => 'home#contacts', as: :contacts
 
   resources :albums
   resources :projects
   resources :images
   resources :photos, only: :destroy
   resources :settings
+
+  root to: 'home#index'
 end
