@@ -1,7 +1,5 @@
 require 'spec_helper'
 
 describe ApplicationController do
-  it 'should use HomeController' do
-    controller.should be_an_instance_of(ApplicationController)
-  end
+  pending
 end
